Merged PR 7304394: Update the nuget packages while working towards changing the feed source
Remove the prototypes folder and move the main solution file to the base directory. Also update the nuget feed to be the artifact exposed in this project, this should allow us to create an azure pipeline to build this project and later produce a nuget package with the interesting binaries of this project.